+++ attest/tpm.go
@@ -16,6 +16,7 @@ package attest
 
 import (
 	"bytes"
+	"crypto"
 	"crypto/rsa"
 	"crypto/sha256"
 	"encoding/base64"
@@ -97,6 +98,21 @@ var (
 		},
 	}
 )
+
+func cryptoHash(h tpm2.Algorithm) (crypto.Hash, error) {
+        switch h {
+        case tpm2.AlgSHA1:
+                return crypto.SHA1, nil
+        case tpm2.AlgSHA256:
+                return crypto.SHA256, nil
+        case tpm2.AlgSHA384:
+                return crypto.SHA384, nil
+        case tpm2.AlgSHA512:
+                return crypto.SHA512, nil
+        default:
+                return crypto.Hash(0), fmt.Errorf("unsupported signature digest: %v", h)
+        }
+}
 
 type tpm20Info struct {
 	vendor       string
