+++ attest/tpm_linux.go
@@ -12,7 +12,7 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-// +build linux
+// +build linux,!gofuzz
 
 package attest
